pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mlp-config"
version = "0.1.0"
description = "Lightweight bootstrap configuration utilities for MLP components."
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
  { name = "Diego" }
]
dependencies = []

[project.optional-dependencies]
dotenv = [
  "python-dotenv>=1.0,<2.0"
]
dev = [
  "pytest>=8.0,<10.0",
  "ruff>=0.8,<1.0",
  "pyright>=1.1,<2.0",
  "build>=1.2,<2.0"
]

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
include = ["mlp*"]
namespaces = true

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pyright]
include = ["src", "tests"]
pythonVersion = "3.11"
typeCheckingMode = "basic"
extraPaths = ["src"]

